Transform the downloaded data, so the close index could be visibly presented + move the corresponding code to "Chart" component
