+++ src/components/stockDashboard/chart/utils.tsx
@@ -1,0 +1,10 @@
+export const convertDate = (date: Date): string =>
+  date.getHours().toLocaleString("en-US", {
+    minimumIntegerDigits: 2,
+    useGrouping: false,
+  }) +
+  ":" +
+  date.getMinutes().toLocaleString("en-US", {
+    minimumIntegerDigits: 2,
+    useGrouping: false,
+  });
