+++ src/components/stockDashboard/models.tsx
@@ -1,0 +1,4 @@
+export interface StockDataModel {
+    time: string;
+    closed: number;
+  }
